cogs pet up pog look at them
